BREAKING zk--parse-file just returns ID and TITLE for single FILE
There is no reason for `zk--parse-file` to do acrobatics to accommodate both
single string filenames and lists of filenames. This syntax allows making it a
fundamental function for accessing ID and TITLE without having to manually call
`string-match`. Since there were only a few calls to it anyway, it was very
easy to update to the new syntax.